Moved the S3 functions from utils.py into S3Utils static methods - updated .gitignore
